% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_honey.R
\name{theme_honey}
\alias{theme_honey}
\title{This is a theme that inspired by my black honey clinique lipstick}
\usage{
theme_honey()
}
\description{
This is a theme that inspired by my black honey clinique lipstick
}
